# prefer nominated node - IMPL