Performance evaluation based on the metric of `scheduler_framework_extension_point_duration_seconds/Filter` in the case of `preemption` when the feature gate is enabled, workloads configuration is 500Nodes, 2000InitPods and 500PodsToSchedule.

xref: https://github.com/kubernetes/kubernetes/pull/96258

indicator  |          before (ms)           |  after  (ms)                      | change
-----------|---------------------------| -------------------------- |---------
Average   | 0.46601329799999947  | 0.3219608919999995     |  -30.9%
Perc50      | 0.2828025477707007    | 0.1457317073170732    |  -48.4%
Perc90      |1.2829493087557604     | 1.0933940774487472    |  -14.8%
Perc99      | 1.4672811059907833    | 1.366742596810934      |  -6.9%


- backup for the legacy approach - component based config and skip the schedule cycle when the nominated node is not fit.
   legacy metric name is `scheduling_algorithm_predicate_evaluation_seconds`, workloads configuration is 500Nodes, 2000InitPods and 500PodsToSchedule.

indicator  |          before (ms)            |  after (ms)                     | change
-----------|---------------------------| --------------------------|---------
Average   |  0.40068560239999973  | 0.20466781999999958 | -48.9%
Perc50      | 0.5357908272610373    | 0.5070993914807302   | -5.3%
Perc90      | 0.9644234890698671     | 0.9127789046653144  | -5.4%
Perc99      | 1.2937062937062935    | 0.9634888438133874   | -25.5%



Signed-off-by: Dave Chen <dave.chen@arm.com>
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

**What type of PR is this?**
 /kind feature


**What this PR does / why we need it**:

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #93013 

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
Overall, enable the feature of `PreferNominatedNode` will  improve the performance of scheduling where preemption might frequently happen, but in theory, enable the feature of `PreferNominatedNode`, the pod might not be scheduled to the best candidate node in the cluster.
```

**Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.**:

Please use the following format for linking documentation:
-  KEP:  https://github.com/kubernetes/enhancements/pull/2026