@chendave 2020-07-17
/cc @Huang-Wei @alculquicondor @ahg-g 

will try to build a benchmark for this later.