@chendave 2020-07-20
https://github.com/kubernetes/kubernetes/blob/43fbe17dc682b703b0b663935a797244378f0aaf/test/integration/scheduler_perf/config/performance-config.yaml#L230-L235 might be the right benchmark needed here, but I found some issue there and currently is still investigating into it.