@Huang-Wei 2020-07-21
@alculquicondor Suppose we go with this new option, should the SchedulerConfig version be bumped to v1beta2 in 1.20?