@alculquicondor 2020-07-21
> Aldo Culquicondor Suppose we go with this new option, should the SchedulerConfig version be bumped to v1beta2 in 1.20?

No. New fields are not a problem.