@chendave 2020-07-26
Have evaluated with `scheduler_perf`,  the result is updated in the cover message, and more testcase is provided.

@Huang-Wei  @alculquicondor  @ahg-g