@alculquicondor 2020-10-28
Note for reviewers: Current plan is to add this feature in 1.21, as per KEP https://github.com/kubernetes/enhancements/tree/master/keps/sig-scheduling/1923-prefer-nominated-node