@chendave 2020-11-06
> Note for reviewers: Current plan is to add this feature in 1.21, as per KEP

I will keep this hold until 1.21 is open.

@alculquicondor  @Huang-Wei  @ahg-g , this is ready for review, I'd highly appreciate if you could take a look before the open of 1.21.