@chendave 2020-12-16
> Please make sure that the KEP is in implementable state first

Thanks, just update the doc is enough? or something else should I do?