@alculquicondor 2020-12-16
Production Readiness Review is now required for 1.21. So you have to add a reviewer for that. But the bar for alpha is not too high.