@alculquicondor 2021-01-18
Will take a look soon, but enhancements freeze is not the same as code freeze, so you are fine.